refactor: 카테고리 서비스 생성
카테고리 서비스 생성과 그에 따른 메소드 분리
